requirements.txt
Pillow==6.0.0
RPi.GPIO==0.6.5
